feat: move `ImputerStrategy` to `safeds.data.tabular.typing`
