+++ src/safeds/data/tabular/transformation/__init__.py
@@ -1,13 +1,12 @@
 """Classes for transforming tabular data."""
 
-from ._imputer import Imputer, ImputerStrategy
+from ._imputer import Imputer
 from ._label_encoder import LabelEncoder
 from ._one_hot_encoder import OneHotEncoder
 from ._table_transformer import InvertibleTableTransformer, TableTransformer
 
 __all__ = [
     "Imputer",
-    "ImputerStrategy",
     "LabelEncoder",
     "OneHotEncoder",
     "InvertibleTableTransformer",
