+++ src/safeds/data/tabular/typing/__init__.py
@@ -2,11 +2,13 @@
 
 from ._column_type import Anything, Boolean, ColumnType, Integer, RealNumber, String
 from ._schema import Schema
+from ._imputer_strategy import ImputerStrategy
 
 __all__ = [
     "Anything",
     "Boolean",
     "ColumnType",
+    "ImputerStrategy",
     "Integer",
     "RealNumber",
     "Schema",
